pyproject.toml
[tool.poetry]
name = "beitech_orders"
version = "0.1.0"
description = ""
authors = ["Santiago Melo Medina"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^1.1.2"
mysql-connector-python = "^8.0.23"

[tool.poetry.dev-dependencies]
Flask-Cors = "^3.0.10"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
